Title: 10개 입력받기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250110/10개 입력받기/receive-10-inputs.c
@@ -16,7 +16,7 @@ int main() {
             cnt++;
         }
     }
-    avg = sum / cnt;
+    avg = (double)sum / cnt;
     printf("%d %.1lf", sum, avg);
 
     return 0;
